osd/PG: fix possible overflow on unfound objects

+++ src/osd/PG.h
@@ -1060,7 +1060,7 @@ public:
   bool have_unfound() const { 
     return missing_loc.have_unfound();
   }
-  int get_num_unfound() const {
+  uint64_t get_num_unfound() const {
     return missing_loc.num_unfound();
   }
